+++ examples/demo.rs
@@ -9,6 +9,7 @@ use resource::resource;
 use winit::{
     event::{ElementState, Event, KeyboardInput, MouseButton, VirtualKeyCode, WindowEvent},
     event_loop::{ControlFlow, EventLoop},
+    window::Window,
 };
 
 mod helpers;
@@ -22,10 +23,7 @@ fn main() {
 }
 
 #[cfg(not(target_arch = "wasm32"))]
-use glutin::PossiblyCurrent;
-
-#[cfg(target_arch = "wasm32")]
-use winit::window::Window;
+use glutin::prelude::*;
 
 pub fn quantize(a: f32, d: f32) -> f32 {
     (a / d + 0.5).trunc() * d
@@ -40,8 +38,9 @@ struct Fonts {
 fn run(
     mut canvas: Canvas<OpenGl>,
     el: EventLoop<()>,
-    #[cfg(not(target_arch = "wasm32"))] windowed_context: glutin::WindowedContext<PossiblyCurrent>,
-    #[cfg(target_arch = "wasm32")] window: Window,
+    #[cfg(not(target_arch = "wasm32"))] context: glutin::context::PossiblyCurrentContext,
+    #[cfg(not(target_arch = "wasm32"))] surface: glutin::surface::Surface<glutin::surface::WindowSurface>,
+    window: Window,
 ) {
     let fonts = Fonts {
         regular: canvas
@@ -106,9 +105,6 @@ fn run(
     let mut perf = PerfGraph::new();
 
     el.run(move |event, _, control_flow| {
-        #[cfg(not(target_arch = "wasm32"))]
-        let window = windowed_context.window();
-
         *control_flow = ControlFlow::Poll;
 
         match event {
@@ -116,7 +112,11 @@ fn run(
             Event::WindowEvent { ref event, .. } => match event {
                 #[cfg(not(target_arch = "wasm32"))]
                 WindowEvent::Resized(physical_size) => {
-                    windowed_context.resize(*physical_size);
+                    surface.resize(
+                        &context,
+                        physical_size.width.try_into().unwrap(),
+                        physical_size.height.try_into().unwrap(),
+                    );
                 }
                 WindowEvent::CursorMoved {
                     device_id: _, position, ..
@@ -309,7 +309,7 @@ fn run(
 
                 canvas.flush();
                 #[cfg(not(target_arch = "wasm32"))]
-                windowed_context.swap_buffers().unwrap();
+                surface.swap_buffers(&context).unwrap();
             }
             Event::MainEventsCleared => window.request_redraw(),
             _ => (),