+++ examples/svg.rs
@@ -4,6 +4,7 @@ use resource::resource;
 use winit::{
     event::{ElementState, Event, KeyboardInput, MouseButton, VirtualKeyCode, WindowEvent},
     event_loop::{ControlFlow, EventLoop},
+    window::Window,
 };
 
 mod helpers;
@@ -17,7 +18,7 @@ fn main() {
 }
 
 #[cfg(not(target_arch = "wasm32"))]
-use glutin::PossiblyCurrent;
+use glutin::prelude::*;
 
 #[cfg(target_arch = "wasm32")]
 use winit::window::Window;
@@ -25,8 +26,9 @@ use winit::window::Window;
 fn run(
     mut canvas: Canvas<OpenGl>,
     el: EventLoop<()>,
-    #[cfg(not(target_arch = "wasm32"))] windowed_context: glutin::WindowedContext<PossiblyCurrent>,
-    #[cfg(target_arch = "wasm32")] window: Window,
+    #[cfg(not(target_arch = "wasm32"))] context: glutin::context::PossiblyCurrentContext,
+    #[cfg(not(target_arch = "wasm32"))] surface: glutin::surface::Surface<glutin::surface::WindowSurface>,
+    window: Window,
 ) {
     canvas
         .add_font_mem(&resource!("examples/assets/Roboto-Light.ttf"))
@@ -62,9 +64,6 @@ fn run(
     println!("Path mem usage: {}kb", total_sisze_bytes / 1024);
 
     el.run(move |event, _, control_flow| {
-        #[cfg(not(target_arch = "wasm32"))]
-        let window = windowed_context.window();
-
         *control_flow = ControlFlow::Poll;
 
         match event {
@@ -72,7 +71,11 @@ fn run(
             Event::WindowEvent { ref event, .. } => match event {
                 #[cfg(not(target_arch = "wasm32"))]
                 WindowEvent::Resized(physical_size) => {
-                    windowed_context.resize(*physical_size);
+                    surface.resize(
+                        &context,
+                        physical_size.width.try_into().unwrap(),
+                        physical_size.height.try_into().unwrap(),
+                    );
                 }
                 WindowEvent::MouseInput {
                     button: MouseButton::Left,
@@ -171,7 +174,7 @@ fn run(
 
                 canvas.flush();
                 #[cfg(not(target_arch = "wasm32"))]
-                windowed_context.swap_buffers().unwrap();
+                surface.swap_buffers(&context).unwrap();
             }
             Event::MainEventsCleared => window.request_redraw(),
             _ => (),